feat!: Rename fsx -> humanfs
fixes #8
