+++ .github/ISSUE_TEMPLATE/change.yml
@@ -8,19 +8,19 @@ body:
       attributes:
           label: Which packages would you like to change?
           options:
-              - label: "`fsx-core`"
+              - label: "`@humanfs/core`"
                 required: false
-              - label: "`fsx-node`"
+              - label: "`@humanfs/node`"
                 required: false
-              - label: "`fsx-deno`"
+              - label: "`@humanfs/deno`"
                 required: false
-              - label: "`fsx-memory`"
+              - label: "`@humanfs/memory`"
                 required: false
-              - label: "`fsx-web`"
+              - label: "`@humanfs/web`"
                 required: false
-              - label: "`fsx-test`"
+              - label: "`@humanfs/test`"
                 required: false
-              - label: "`fsx-types`"
+              - label: "`@humanfs/types`"
                 required: false
     - type: textarea
       attributes:
@@ -28,16 +28,16 @@ body:
           description: |
               Please explain your use case in as much detail as possible without referring to a specific solution.
           placeholder: |
-              fsx currently...
+              humanfs currently...
       validations:
           required: true
     - type: textarea
       attributes:
           label: What do you think is the correct solution?
           description: |
-              Please explain how you'd like to change fsx to address the problem.
+              Please explain how you'd like to change humanfs to address the problem.
           placeholder: |
-              I'd like fsx to...
+              I'd like humanfs to...
       validations:
           required: true
     - type: checkboxes
