+++ README.md
@@ -1,4 +1,4 @@
-# fsx: A modern filesystem API for JavaScript
+# humanfs: A modern filesystem API for JavaScript
 
 by [Nicholas C. Zakas](https://humanwhocodes.com)
 
@@ -6,13 +6,13 @@ If you find this useful, please consider supporting my work with a [donation](ht
 
 ## Description
 
-This is the monorepo for fsx, a modern filesystem API for JavaScript. Most of the filesystem APIs provided by JavaScript runtimes were designed to emulate Linux utilities which, while making them easy to adopt and understand, often requires writing way more code than necessary for common tasks. fsx is a new approach that streamlines the most common operations while providing useful ways to test the expected functionality.
+This is the monorepo for humanfs, a modern filesystem API for JavaScript. Most of the filesystem APIs provided by JavaScript runtimes were designed to emulate Linux utilities which, while making them easy to adopt and understand, often requires writing way more code than necessary for common tasks. humanfs is a new approach that streamlines the most common operations while providing useful ways to test the expected functionality.
 
-## Get Started Using fsx
+## Get Started Using humanfs
 
-To get started using fsx, please see the [Getting Started](./docs/README.md).
+To get started using humanfs, please see the [Getting Started](./docs/README.md).
 
-## Contribute to fsx
+## Contribute to humanfs
 
 See the [Contributing Guide](./docs/CONTRIBUTING.md).
