+++ docs/file-operations.md
@@ -4,25 +4,25 @@
 
 Files are read by using the method that returns the specific data type that you want:
 
--   `fsx.text(filePath)` reads the given file and returns a string.
--   `fsx.json(filePath)` reads the given file and returns a JSON value.
--   `fsx.bytes(filePath)` reads the given file and returns an `Uint8Array`.
--   **Deprecated:** `fsx.arrayBuffer(filePath)` reads the given file and returns an `ArrayBuffer`.
+-   `hfs.text(filePath)` reads the given file and returns a string.
+-   `hfs.json(filePath)` reads the given file and returns a JSON value.
+-   `hfs.bytes(filePath)` reads the given file and returns an `Uint8Array`.
+-   **Deprecated:** `hfs.arrayBuffer(filePath)` reads the given file and returns an `ArrayBuffer`.
 
 Here are some examples:
 
 ```js
 // read plain text
-const text = await fsx.text("/path/to/file.txt");
+const text = await hfs.text("/path/to/file.txt");
 
 // read JSON
-const json = await fsx.json("/path/to/file.json");
+const json = await hfs.json("/path/to/file.json");
 
 // read bytes
-const bytes = await fsx.bytes("/path/to/file.png");
+const bytes = await hfs.bytes("/path/to/file.png");
 
 // read buffer
-const buffer = await fsx.arrayBuffer("/path/to/file.png");
+const buffer = await hfs.arrayBuffer("/path/to/file.png");
 ```
 
 ### When Files Don't Exist
@@ -30,7 +30,7 @@ const buffer = await fsx.arrayBuffer("/path/to/file.png");
 When you attempt to read a file that doesn't exist, each of these methods returns `undefined`. This is different than most filesystem APIs, which typically throw an error when a file doesn't exist. This allows you to avoid wrapping each call in `try-catch` just to account for files that don't exist. Instead, you can use an `if` statement:
 
 ```js
-const text = await fsx.text("/path/to/file.txt");
+const text = await hfs.text("/path/to/file.txt");
 
 if (!text) {
 	// handle when the file doesn't exist
@@ -41,19 +41,19 @@ Alternatively, you can also use the nullish coalescing operator (`??`) to specif
 
 ```js
 // read plain text
-const text = (await fsx.text("/path/to/file.txt")) ?? "default value";
+const text = (await hfs.text("/path/to/file.txt")) ?? "default value";
 
 // read JSON
-const json = (await fsx.json("/path/to/file.json")) ?? {};
+const json = (await hfs.json("/path/to/file.json")) ?? {};
 
 // read bytes
 const bytes =
-	(await fsx.bytes("/path/to/file.png")) ?? new Uint8Array();
+	(await hfs.bytes("/path/to/file.png")) ?? new Uint8Array();
 ```
 
 ## Writing Files
 
-To write files, call the `fsx.write()` method. This method accepts two arguments:
+To write files, call the `hfs.write()` method. This method accepts two arguments:
 
 -   `filePath:string` - the path to write to
 -   `value:string|ArrayBuffer|ArrayBufferView` - the value to write to the file
@@ -61,11 +61,11 @@ To write files, call the `fsx.write()` method. This method accepts two arguments
 Here's an example:
 
 ```js
-await fsx.write("/path/to/file.txt", "Hello world!");
+await hfs.write("/path/to/file.txt", "Hello world!");
 
 const bytes = new TextEncoder().encode("Hello world!");
 
-await fsx.write("/path/to/file.txt", bytes);
+await hfs.write("/path/to/file.txt", bytes);
 ```
 
 > [!TIP]
@@ -73,29 +73,29 @@ await fsx.write("/path/to/file.txt", bytes);
 
 ## Detecting Files
 
-To determine to if a file exists, use the `fsx.isFile(filePath)` method, which returns `true` if the given file exists or `false` otherwise.
+To determine to if a file exists, use the `hfs.isFile(filePath)` method, which returns `true` if the given file exists or `false` otherwise.
 
 ```js
-if (await fsx.isFile("/path/to/file.txt")) {
+if (await hfs.isFile("/path/to/file.txt")) {
 	// handle the file
 }
 ```
 
 > [!IMPORTANT]
-> If the file doesn't exist, `fsx.isFile()` returns `false`. This method does not throw errors unless the filesystem cannot be accessed.
+> If the file doesn't exist, `hfs.isFile()` returns `false`. This method does not throw errors unless the filesystem cannot be accessed.
 
 ## Deleting Files
 
-To delete files, call the `fsx.delete(filePath)` method. For example:
+To delete files, call the `hfs.delete(filePath)` method. For example:
 
 ```js
-await fsx.delete("/path/to/file.txt");
+await hfs.delete("/path/to/file.txt");
 ```
 
 ## Retrieving File Size
 
-To get the size of a file in bytes, call the `fsx.size(filePath)` method. This method returns the size in bytes of the file if found and `undefined` if not found. Here's an example:
+To get the size of a file in bytes, call the `hfs.size(filePath)` method. This method returns the size in bytes of the file if found and `undefined` if not found. Here's an example:
 
 ```js
-const size = await fsx.size("/path/to/file.txt");
+const size = await hfs.size("/path/to/file.txt");
 ```
