+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "fsx",
+  "name": "hfs",
   "version": "1.0.0",
   "private": true,
   "type": "module",
