+++ packages/core/README.md
@@ -1,4 +1,4 @@
-# `fsx-core`
+# `@humanfs/core`
 
 by [Nicholas C. Zakas](https://humanwhocodes.com)
 
@@ -6,9 +6,9 @@ If you find this useful, please consider supporting my work with a [donation](ht
 
 ## Description
 
-The core functionality for `fsx` that is shared across all implementations for all runtimes. The contents of this package are intentionally runtime agnostic and are not intended to be used alone.
+The core functionality for `hfs` that is shared across all implementations for all runtimes. The contents of this package are intentionally runtime agnostic and are not intended to be used alone.
 
-Currently, this package simply exports the `Fsx` class, which is an abstract base class intended to be inherited from in runtime-specific fsx packages (like `fsx-node`).
+Currently, this package simply exports the `Hfs` class, which is an abstract base class intended to be inherited from in runtime-specific hfs packages (like `@humanfs/node`).
 
 > [!WARNING]
 > This project is **experimental** and may change significantly before v1.0.0. Use at your own caution and definitely not in production!
@@ -20,25 +20,25 @@ Currently, this package simply exports the `Fsx` class, which is an abstract bas
 Install using your favorite package manager for Node.js:
 
 ```shell
-npm install fsx-core
+npm install @humanfs/core
 
 # or
 
-pnpm install fsx-core
+pnpm install @humanfs/core
 
 # or
 
-yarn add fsx-core
+yarn add @humanfs/core
 
 # or
 
-bun install fsx-core
+bun install @humanfs/core
 ```
 
-Then you can import the `Fsx` class like this:
+Then you can import the `Hfs` class like this:
 
 ```js
-import { Fsx } from "fsx-core";
+import { Hfs } from "@humanfs/core";
 ```
 
 ### Deno
@@ -48,15 +48,15 @@ For Deno, set up a `deno.json` file like this:
 ```json
 {
 	"imports": {
-		"fsx-core": "npm:fsx-core@latest"
+		"@humanfs/core": "npm:@humanfs/core@latest"
 	}
 }
 ```
 
-Then you can import the `Fsx` class like this:
+Then you can import the `Hfs` class like this:
 
 ```js
-import { Fsx } from "fsx-core";
+import { Hfs } from "@humanfs/core";
 ```
 
 ### Browser
@@ -64,31 +64,31 @@ import { Fsx } from "fsx-core";
 It's recommended to import the minified version to save bandwidth:
 
 ```js
-import { Fsx } from "https://cdn.skypack.dev/fsx-core?min";
+import { Hfs } from "https://cdn.skypack.dev/@humanfs/core?min";
 ```
 
 However, you can also import the unminified version for debugging purposes:
 
 ```js
-import { Fsx } from "https://cdn.skypack.dev/fsx-core";
+import { Hfs } from "https://cdn.skypack.dev/@humanfs/core";
 ```
 
 ## Usage
 
-The `Fsx` class contains all of the basic functionality for an `Fsx` instance *without* a predefined impl. This class is mostly used for creating runtime-specific impls, such as `NodeFsx` and `DenoFsx`.
+The `Hfs` class contains all of the basic functionality for an `Hfs` instance *without* a predefined impl. This class is mostly used for creating runtime-specific impls, such as `NodeHfs` and `DenoHfs`.
 
 You can create your own instance by providing an `impl` directly:
 
 ```js
-const fsx = new Fsx({ impl: { async text() {} }});
+const hfs = new Hfs({ impl: { async text() {} }});
 ```
 
 The specified `impl` becomes the base impl for the instance, meaning you can always reset back to it using `resetImpl()`.
 
-You can also inherit from `Fsx` to create your own class with a preconfigured impl, such as:
+You can also inherit from `Hfs` to create your own class with a preconfigured impl, such as:
 
 ```js
-class MyFsx extends Fsx {
+class MyHfs extends Hfs {
 	constructor() {
 		super({
 			impl: myImpl
