+++ packages/core/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "fsx-core",
+  "name": "@humanfs/core",
   "version": "0.5.0",
-  "description": "The core of the fsx library.",
+  "description": "The core of the humanfs library.",
   "type": "module",
   "types": "dist/index.d.ts",
   "exports": {
@@ -22,22 +22,25 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/humanwhocodes/fsx.git"
+    "url": "git+https://github.com/humanwhocodes/humanfs.git"
+  },
+  "publishConfig": {
+    "access": "public"
   },
   "keywords": [
     "filesystem",
     "fs",
-    "fsx",
+    "hfs",
     "files"
   ],
   "author": "Nicholas C. Zakas",
   "license": "Apache-2.0",
   "bugs": {
-    "url": "https://github.com/humanwhocodes/fsx/issues"
+    "url": "https://github.com/humanwhocodes/humanfs/issues"
   },
-  "homepage": "https://github.com/humanwhocodes/fsx#readme",
+  "homepage": "https://github.com/humanwhocodes/humanfs#readme",
   "devDependencies": {
-    "fsx-types": "^0.4.0",
+    "@humanfs/types": "^0.4.0",
     "c8": "^9.0.0",
     "mocha": "^10.2.0",
     "typescript": "^5.2.2"
