+++ packages/memory/package.json
@@ -1,13 +1,13 @@
 {
-  "name": "fsx-memory",
+  "name": "@humanfs/memory",
   "version": "0.5.0",
-  "description": "The in-memory bindings for the fsx library.",
+  "description": "The in-memory bindings for the hfs library.",
   "type": "module",
-  "types": "dist/memory-fsx.d.ts",
+  "types": "dist/memory-hfs.d.ts",
   "exports": {
     "import": {
-      "default": "./src/memory-fsx.js",
-      "types": "./dist/memory-fsx.d.ts"
+      "default": "./src/memory-hfs.js",
+      "types": "./dist/memory-hfs.d.ts"
     }
   },
   "files": [
@@ -22,28 +22,31 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/humanwhocodes/fsx.git"
+    "url": "git+https://github.com/humanwhocodes/humanfs.git"
+  },
+  "publishConfig": {
+    "access": "public"
   },
   "keywords": [
     "filesystem",
     "fs",
-    "fsx",
+    "hfs",
     "files"
   ],
   "author": "Nicholas C. Zakas",
   "license": "Apache-2.0",
   "bugs": {
-    "url": "https://github.com/humanwhocodes/fsx/issues"
+    "url": "https://github.com/humanwhocodes/humanfs/issues"
   },
-  "homepage": "https://github.com/humanwhocodes/fsx#readme",
+  "homepage": "https://github.com/humanwhocodes/humanfs#readme",
   "devDependencies": {
     "@types/node": "^20.9.4",
     "mocha": "^10.2.0",
     "typescript": "^5.2.2",
-    "fsx-test": "^0.5.0",
-    "fsx-types": "^0.4.0"
+    "@humanfs/test": "^0.5.0",
+    "@humanfs/types": "^0.4.0"
   },
   "dependencies": {
-    "fsx-core": "^0.5.0"
+    "@humanfs/core": "^0.5.0"
   }
 }
