+++ packages/memory/tests/memory-hfs.test.js
@@ -1,5 +1,5 @@
 /**
- * @fileoverview Tests for the Fsx class.
+ * @fileoverview Tests for the Hfs class.
  * @author Nicholas C. Zakas
  */
 
@@ -7,8 +7,8 @@
 // Imports
 //------------------------------------------------------------------------------
 
-import { MemoryFsxImpl } from "../src/memory-fsx.js";
-import { FsxImplTester } from "fsx-test";
+import { MemoryHfsImpl } from "../src/memory-hfs.js";
+import { HfsImplTester } from "@humanfs/test";
 import assert from "node:assert";
 
 //------------------------------------------------------------------------------
@@ -22,13 +22,13 @@ const fixturesDir = "fixtures";
 // Tests
 //------------------------------------------------------------------------------
 
-const tester = new FsxImplTester({
+const tester = new HfsImplTester({
 	outputDir: fixturesDir,
 	assert,
 	test: globalThis,
 });
 
 await tester.test({
-	name: "MemoryFsxImpl",
-	impl: new MemoryFsxImpl({ volume }),
+	name: "MemoryHfsImpl",
+	impl: new MemoryHfsImpl({ volume }),
 });
