+++ packages/memory/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "files": ["src/memory-fsx.js"],
+  "files": ["src/memory-hfs.js"],
   "compilerOptions": {
     "declaration": true,
     "emitDeclarationOnly": true,
