+++ packages/node/src/node-hfs.js
@@ -1,5 +1,5 @@
 /**
- * @fileoverview The main file for the fsx package.
+ * @fileoverview The main file for the hfs package.
  * @author Nicholas C. Zakas
  */
 /* global Buffer:readonly */
@@ -8,8 +8,8 @@
 // Types
 //-----------------------------------------------------------------------------
 
-/** @typedef {import("fsx-types").FsxImpl} FsxImpl */
-/** @typedef {import("fsx-types").FsxDirectoryEntry} FsxDirectoryEntry */
+/** @typedef {import("@humanfs/types").HfsImpl} HfsImpl */
+/** @typedef {import("@humanfs/types").HfsDirectoryEntry} HfsDirectoryEntry */
 /** @typedef {import("node:fs/promises")} Fsp */
 /** @typedef {import("fs").Dirent} Dirent */
 
@@ -17,7 +17,7 @@
 // Imports
 //-----------------------------------------------------------------------------
 
-import { Fsx } from "fsx-core";
+import { Hfs } from "@humanfs/core";
 import path from "node:path";
 import { Retrier } from "@humanwhocodes/retry";
 import nativeFsp from "node:fs/promises";
@@ -34,9 +34,9 @@ const RETRY_ERROR_CODES = new Set(["ENFILE", "EMFILE"]);
 
 /**
  * A class representing a directory entry.
- * @implements {FsxDirectoryEntry}
+ * @implements {HfsDirectoryEntry}
  */
-class NodeFsxDirectoryEntry {
+class NodeHfsDirectoryEntry {
 	/**
 	 * The name of the directory entry.
 	 * @type {string}
@@ -78,10 +78,10 @@ class NodeFsxDirectoryEntry {
 //-----------------------------------------------------------------------------
 
 /**
- * A class representing the Node.js implementation of Fsx.
- * @implements {FsxImpl}
+ * A class representing the Node.js implementation of Hfs.
+ * @implements {HfsImpl}
  */
-export class NodeFsxImpl {
+export class NodeHfsImpl {
 	/**
 	 * The file system module to use.
 	 * @type {Fsp}
@@ -304,7 +304,7 @@ export class NodeFsxImpl {
 	/**
 	 * Returns a list of directory entries for the given path.
 	 * @param {string} dirPath The path to the directory to read.
-	 * @returns {AsyncIterable<FsxDirectoryEntry>} A promise that resolves with the
+	 * @returns {AsyncIterable<HfsDirectoryEntry>} A promise that resolves with the
 	 *   directory entries.
 	 * @throws {TypeError} If the directory path is not a string.
 	 * @throws {Error} If the directory cannot be read.
@@ -315,7 +315,7 @@ export class NodeFsxImpl {
 		});
 
 		for (const entry of entries) {
-			yield new NodeFsxDirectoryEntry(entry);
+			yield new NodeHfsDirectoryEntry(entry);
 		}
 	}
 
@@ -341,17 +341,17 @@ export class NodeFsxImpl {
 
 /**
  * A class representing a file system utility library.
- * @implements {FsxImpl}
+ * @implements {HfsImpl}
  */
-export class NodeFsx extends Fsx {
+export class NodeHfs extends Hfs {
 	/**
 	 * Creates a new instance.
 	 * @param {object} [options] The options for the instance.
 	 * @param {Fsp} [options.fsp] The file system module to use.
 	 */
 	constructor({ fsp } = {}) {
-		super({ impl: new NodeFsxImpl({ fsp }) });
+		super({ impl: new NodeHfsImpl({ fsp }) });
 	}
 }
 
-export const fsx = new NodeFsx();
+export const hfs = new NodeHfs();
