+++ packages/test/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "fsx-test",
+  "name": "@humanfs/test",
   "version": "0.5.0",
-  "description": "Testing utilities for fsx",
+  "description": "Testing utilities for hfs",
   "type": "module",
   "types": "dist/index.d.ts",
   "exports": {
@@ -20,20 +20,23 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/humanwhocodes/fsx.git"
+    "url": "git+https://github.com/humanwhocodes/humanfs.git"
+  },
+  "publishConfig": {
+    "access": "public"
   },
   "keywords": [
     "filesystem",
     "fs",
-    "fsx",
+    "hfs",
     "files"
   ],
   "author": "Nicholas C. Zakas",
   "license": "Apache-2.0",
   "bugs": {
-    "url": "https://github.com/humanwhocodes/fsx/issues"
+    "url": "https://github.com/humanwhocodes/humanfs/issues"
   },
-  "homepage": "https://github.com/humanwhocodes/fsx#readme",
+  "homepage": "https://github.com/humanwhocodes/humanfs#readme",
   "devDependencies": {
     "typescript": "^5.2.2"
   }
