+++ packages/web/package.json
@@ -1,14 +1,14 @@
 {
-  "name": "fsx-web",
+  "name": "@humanfs/web",
   "version": "0.1.0",
-  "description": "The in-web bindings for the fsx library.",
+  "description": "The in-web bindings for the hfs library.",
   "type": "module",
-  "types": "dist/web-fsx.d.ts",
-  "main": "dist/web-fsx.js",
+  "types": "dist/web-hfs.d.ts",
+  "main": "dist/web-hfs.js",
   "exports": {
     "import": {
-      "default": "./src/web-fsx.js",
-      "types": "./dist/web-fsx.d.ts"
+      "default": "./src/web-hfs.js",
+      "types": "./dist/web-hfs.d.ts"
     }
   },
   "files": [
@@ -23,12 +23,15 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/humanwhocodes/fsx.git"
+    "url": "git+https://github.com/humanwhocodes/humanfs.git"
+  },
+  "publishConfig": {
+    "access": "public"
   },
   "keywords": [
     "filesystem",
     "fs",
-    "fsx",
+    "hfs",
     "files",
     "opfs",
     "web"
@@ -36,21 +39,21 @@
   "author": "Nicholas C. Zakas",
   "license": "Apache-2.0",
   "bugs": {
-    "url": "https://github.com/humanwhocodes/fsx/issues"
+    "url": "https://github.com/humanwhocodes/humanfs/issues"
   },
-  "homepage": "https://github.com/humanwhocodes/fsx#readme",
+  "homepage": "https://github.com/humanwhocodes/humanfs#readme",
   "devDependencies": {
     "@wdio/browser-runner": "^8.29.1",
     "@wdio/cli": "^8.29.1",
     "@wdio/mocha-framework": "^8.29.1",
     "@wdio/spec-reporter": "^8.29.1",
     "chai": "^5.0.0",
-    "fsx-test": "^0.5.0",
-    "fsx-types": "^0.4.0",
+    "@humanfs/test": "^0.5.0",
+    "@humanfs/types": "^0.4.0",
     "mocha": "^10.2.0",
     "typescript": "^5.2.2"
   },
   "dependencies": {
-    "fsx-core": "^0.5.0"
+    "@humanfs/core": "^0.5.0"
   }
 }
